feat: switch to better-tailwindcss eslint plugin (#71)
since `eslint-plugin-tailwindcss` is no longer actively maintained, and
the `alpha` fork we are currently using does not seem to move forward,
we should switch to `eslint-plugin-better-tailwindcss`.

EDIT: there has been a recent beta release of
`eslint-plugin-tailwindcss` with v4 support (which re-publishes the
alpha fork we are currently using), but it is unclear how long a stable
release will take, because they are attempting a full rewrite of the
plugin.